requirements.txt
hikari-lightbulb
hikari-miru
asyncpg
timeago
python-dotenv
lightbulb-ext-filament